+++ tamr_client/attribute/_attribute.py
@@ -3,7 +3,7 @@ See https://docs.tamr.com/reference/attribute-types
 """
 from copy import deepcopy
 from dataclasses import replace
-from typing import Optional, Tuple
+from typing import Optional
 
 from tamr_client import response
 from tamr_client._types import Attribute, AttributeType, Dataset, JsonDict, Session, URL
@@ -103,31 +103,6 @@ def _from_json(url: URL, data: JsonDict) -> Attribute:
         is_nullable=cp["isNullable"],
         type=attribute_type.from_json(cp["type"]),
     )
-
-
-def from_dataset_all(session: Session, dataset: Dataset) -> Tuple[Attribute, ...]:
-    """Get all attributes from a dataset
-
-    Args:
-        dataset: Dataset containing the desired attributes
-
-    Returns:
-        The attributes for the specified dataset
-
-    Raises:
-        requests.HTTPError: If an HTTP error is encountered.
-    """
-    attrs_url = replace(dataset.url, path=dataset.url.path + "/attributes")
-    r = session.get(str(attrs_url))
-    attrs_json = response.successful(r).json()
-
-    attrs = []
-    for attr_json in attrs_json:
-        id = attr_json["name"]
-        attr_url = replace(attrs_url, path=attrs_url.path + f"/{id}")
-        attr = _from_json(attr_url, attr_json)
-        attrs.append(attr)
-    return tuple(attrs)
 
 
 def to_json(attr: Attribute) -> JsonDict: